Added benchmark that logs memory for component class bitmasks
