+++ package.json
@@ -29,15 +29,15 @@
     "@pika/plugin-ts-standard-pkg": "^0.6.0",
     "@types/benchmark": "^1.0.31",
     "@types/jest": "^24.0.17",
-    "@types/lodash": "^4.14.136",
-    "@types/node": "^12.7.1",
+    "@types/lodash": "^4.14.137",
+    "@types/node": "^12.7.2",
     "benchmark": "^2.1.4",
-    "husky": "^3.0.3",
-    "jest": "^24.8.0",
+    "husky": "^3.0.4",
+    "jest": "^24.9.0",
     "lodash": "^4.17.15",
     "nodemon": "^1.19.0",
     "prettier": "^1.17.0",
-    "rimraf": "^2.6.3",
+    "rimraf": "^3.0.0",
     "ts-jest": "^24.0.2",
     "ts-node": "^8.3.0",
     "tslint": "^5.16.0",
